Updating to Last Match State
So…apparently I never actually committed the code after World's to GitHub, but I still had it on my flash drive… This is the state of our code at the last match. (I made sure to not delete your comments, Daniel.)
This stuff shouldn't really change any of the main robot stuff; it just accounts for variable values and auto plans that were constantly tweaked during the competition.