Commented Auto Structure Classes
These are the most important of the Auto classes. These involve how autos are constructed of multiple plans (in AutoBuilder), where the field layout is received (AutoRunner), which plan gets chosen to be run (AutoRunner), how the auto is terminated when auto mode ends (AutoRunner), and provides a place for the most common driving and rotating methods that all auto plans will inherit (AutoPlan).