Commented Auto Movement Classes
These are the brunt of testing and fine-tuning auto work. Once these are good, you can generally just work on your plans, but getting basic driving in a line and rotating to work accurately and quickly is a hard task. My methods here are much more complicated than the ones we wrote in our first year of auto work, so you can get by with much simpler calculations if you make your own Drive and Rotate classes.

There's no point commenting the actual auto plans, because those are specific to last year's game and are just a list of commands to get from place to place. A lot of testing is just tweaking distances and angles on those, but they're easy to understand. You'll be making your own plans for next game, anyway.